commented and cleaned code

+++ API/meetingAPI.py
@@ -1,74 +1,84 @@
+# API imports
 from fastapi import FastAPI, Response, status, HTTPException, Request, Body
 import uvicorn
 from pydantic import BaseModel
+
+# for type annotations
 from typing import Any, List, Mapping
 
+# imports to be able to use functions from other modules in project or access files
 import sys
 from pathlib import Path
 sys.path.append(str(Path(__file__).resolve().parent.parent))
 import os
 
-from tools.get_port import get_free_port
+# functions in project
+from tools.get_port import get_free_port # for finding a free port for the proxy
 from tools.json_to_scribble import transform
-from scribble_python.wf_checker import check_well_formedness, WellFormednessError
+from scribble_python.wf_checker import check_well_formedness, WellFormednessError # check Scribble protocol
+from proxy.session_logic.type_validation import create_type_checker # for payload type schemas
 
-# fro handling several proxies
-import threading
+# proxy imports
+import threading # for handling several proxies
 from proxy.proxy import main_proxy
 
-from proxy.session_logic.type_validation import create_type_checker
-
-# -- API ------------------------------------------------------
+# -- initialize ----------------------------------------------------------------------
 app = FastAPI()
 
+# models for receiving info through API requests
 class Role(BaseModel):
     name: str
     alias: str
 
 class Meeting(BaseModel):
-    protocol: str
-    roles: List[Role]
-    body: List[Any]  # steps, for now
-
+    protocol: str # name of protocol
+    roles: List[Role] # actors in protocol
+    body: List[Any]  # protocol definition
 
 # dicts for storing meetings info
-meetingDict: dict[str, Meeting] = {}
-meeting_info: dict[str, int] = {}
+meetingDict: dict[str, Meeting] = {} # name of meeting : meeting object
+meeting_info: dict[str, int] = {} # name of meeting : assigned proxy port
 current_json:dict = None  # to store the current json protocol
-meeting_types:dict[str, list] = {} # meeting name: schema list 
-
+meeting_types:dict[str, list] = {} # meeting name: schema list
 # for handling several proxies
-proxy_threads = {}  # meeting name -> thread
+proxy_threads = {}  # meeting name : thread
 
+# -- API ---------------------------------------------------------------------------------------------------
 
-# create a meeting -> register and give back meeting ID
+# create a meeting -> register and give back corresponding proxy port
 @app.post("/meetings/", response_model=int)
-async def createMeetingReq(meeting: Meeting, request: Request) -> int: # why Any?
+async def createMeetingReq(meeting: Meeting, request: Request) -> int:
+    # TODO: check this error works
+    if meeting.protocol in meetingDict:
+        raise HTTPException(status_code=409, detail="Meeting already exists")
+
     # zero: define types
-    print(f"meeting protocol: {meeting.protocol}") # debug
+    print(f"defning meeting protocol: {meeting.protocol}") # debug
     if meeting.protocol in meeting_types:
-        print(f"found schemas in {meeting.protocol}") # debug
+        # print(f"found schemas in {meeting.protocol}") # debug
         schemas = meeting_types[meeting.protocol]
     else:
         schemas = None
-    types = create_type_checker(schemas)
+    types = create_type_checker(schemas) # all type schemas we'll need for types in protocol
 
     # one: transform to scr
     current_json = await request.json()
     transform(current_json, types, output_dir="API/protocols")
+
     # two: check if protocol is well-formed
     try:
         check_well_formedness(f"API/protocols/{meeting.protocol}.scr")
     except WellFormednessError as e:
         os.remove(f"API/protocols/{meeting.protocol}.scr")
         return Response(status_code=status.HTTP_400_BAD_REQUEST, content=f"Invalid protocol: {e}")
-    # three: find free port
+    
+    # three: find free port for proxy
     proxy_port = get_free_port()
-    # four: store meeting info and schemas
+
+    # four: store meeting info
     meetingDict[meeting.protocol] = meeting
     proxy_port = get_free_port()
     meeting_info[meeting.protocol] = proxy_port
-
     
     # five: create proxy for meeting
     actors = [(role.name, role.alias) for role in meeting.roles]
@@ -78,25 +88,40 @@ async def createMeetingReq(meeting: Meeting, request: Request) -> int: # why Any
     thread = threading.Thread(target=run_proxy, daemon=True)
     thread.start()
     proxy_threads[meeting.protocol] = thread
+
     # six: give back port number
     return proxy_port
 
 # so that others can find out port based on meeting name
-@app.get("/meetings/{meeting_id}/", response_model=int)
-async def getMeetingReq(name:str) -> int:
-    if name not in meetingDict:
+@app.get("/meetings/{meeting_name}/", response_model=int)
+async def getMeetingReq(meeting_name:str) -> int:
+    if meeting_name not in meetingDict:
         raise HTTPException(404, "Not found")
-    return meeting_info[name]
+    return meeting_info[meeting_name]
 
-# for now schemas one by one but later as a batch
+# allows definition of JSON schemas for types used in a meeting
+# TODO:for now schemas one by one but later as a batch
 @app.post("/types/{meeting_name}/")
-async def createMeetingReq(meeting_name: str, schema:Mapping[str, Any]= Body(...)): # why Any?
-    print(f"defining: {schema} for {meeting_name}") # debug
-    if meeting_name not in meeting_types:
+async def createMeetingReq(meeting_name: str, schema:Mapping[str, Any]= Body(...)): # TODO: why Any?
+    print(f"defining: {schema.get('title')} for {meeting_name}") # debug
+    if meeting_name not in meeting_types: # for first custom schema in a meeting
         meeting_types[meeting_name] = []
     meeting_types[meeting_name].append(schema)
 
+# for deleting a meeting once it is done
+# TODO: use in proxies to close them off
+@app.delete("/meetings/{meeting_name}")
+async def deleteMeetingReq(meeting_name: str):
+    if meeting_name not in meetingDict:
+        raise HTTPException(404, "Not found")
+    print(f"deleting meeting {meeting_name}") # debug
+    del meetingDict[meeting_name]
+    del meeting_info[meeting_name]
+    del meeting_types[meeting_name]
+    return Response(status_code=status.HTTP_204_NO_CONTENT) # TODO: check this is the right thing to return
+    # should return ok but idk how it works
 
+# TODO: determine if this is necessary
 """
 @app.get("/meetings/{meeting_id}/{actors}", response_model=str)
 async def getMeetingReq(meeting_id:str) -> str:
@@ -105,11 +130,3 @@ async def getMeetingReq(meeting_id:str) -> str:
     return "" # temporary; return local protocol projection
 """
 
-@app.delete("/meetings/{meeting_id}")
-async def deleteMeetingReq(meeting_id: str) -> Response:
-    if meeting_id not in meetingDict:
-        raise HTTPException(404, "Not found")
-    del meetingDict[meeting_id]
-    del meeting_info[meeting_id]
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
-    # should return ok but idk how it works
